requirements.txt
python-telegram-bot==20.7
opencv-python-headless==4.8.1.78
numpy==1.24.3
python-dotenv==1.0.0
jwt==1.4.0
requests==2.31.0
ultralytics==8.3.235
